requirements.txt
flask
flask_restful
requests
evolved5g
flask_sqlalchemy
flask_migrate
flask_marshmallow
marshmallow-sqlalchemy
python-dotenv